requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
email-validator
python-multipart
psycopg2-binary
motor